tests/main/snapd-snap: Remove snapd snap on cleanup
Use find instead of ls -1 to count the number of installed snapd snap
revisions so we can only revert snapd on cleanup if there is more than 1
revision of it installed. Otherwise we remove the snapd snap which should
then restore the spread system back to it's prior state before we installed
the snapd snap.