many: wait for up to 10min for NTP syncronization before autorefresh
We got bugreports that on systems without a proper clock the
auto-refresh code gets confused and reports strange refresh times.

It looks like this is triggered by snapd starting up and while in
the middle of the startup the NTP sync happens and that confuses
the autorefresh code that tries to calculate the next refresh window
because the clock is jumping.

To avoid this happening this commit adds code that waits for up
to 10min for the NTP syncronization to happen. This does not
fully eliminate the bug (that we still don't fully understand)
but it should fix it for most systems that have some sort of
network connection.