cmd/snap-failure: use snapd from the snapd snap if core is not present
When there is no previous revision of snapd, the code would fall back to start
snapd binary from the current core version. However, it is possible that the
core snap is not present in the system at all. In which case we should fall back
to the current snapd snap.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>