tests/lib/nested.sh: allow disabling repacking of fundamental snaps too
When using NESTED_BUILD_SNAPD_FROM_CURRENT, it is also sometimes desirable to
be able to disable repacking of the other snaps too so that we can just repack
the snapd snap and keep all the other snaps from upstream as-is.

Add some variables to nested.sh to enable this.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>