tests/main/snapd-snap: Restore packaging and set version for trusty
prepare-restore builds the deb and hence sets the version in
debian/changelog before the test normally runs. But since on trusty we go
and swap out the debian/ directory with the xenial one so it can build
properly in LXD, we also need to do the version bump, otherwise the built
snap has a version less than the deb we are already running and so will not
get reexec'd into since its version is not greater. So steal the version
bumping logic from prepare-restore.

Also make sure we restore back the trusty specific packaging in case this
is still needed by other subsequent tests.