New variables used to create nested images using different channels
The idea of this change is to define a generic en var called
NESTED_CHANNEL which is used as main channel when the image is created,
and anotherone called NESTED_REFRESH_CHANNEL to refresh. But also there
are new variables used for other snaps.

NESTED_KERNEL_CHANNEL: Channel used for the kernel snap
NESTED_BASE_CHANNEL: Channel used for the base snap
NESTED_GADGET_CHANNEL: Channel used for the gadget snap
NESTED_SNAPD_CHANNEL: Channel used for the snapd snap