o/snapstate: move unlock/lock into tests' settle
This commit moves the unlock and lock around settle calls into the settle
function. This ensures that a failure in settle won't cause the test to
end by unlocking already unlocked state (causing a panic) and it doesn't
clutter tests with two pairs of (un)locks and their deferred opposites.