tests/core/snapd-refresh-vs-services-reboots: use core18 from stable
This test has been seen failing several times in ubuntu-core-18 with
this lines in the logs:

    systemd[1]: Dependency failed for Mount unit for test-snapd-svc-flip-flop, revision x1.
    systemd[1]: Dependency failed for Service for snap application test-snapd-svc-flip-flop.svc1.

(more complete logs are here: https://paste.ubuntu.com/p/r7nyM4krH4/)
It looks like a failure in systemd, which deactivates our snap's mount
unit when it gets reloaded just before starting the service contained in
the snap. This doesn't seem to happen in the stable channel of core18,
so let's use that one, while we investigate the issue.