tests/core20-new-snapd-does-not-break-old-initrd: avoid auto-refreshes in test
Avoid the problem of snapd being auto-refreshed during first boot. We sometimes
see this issue because when we build the image, we end up with a snapd snap
from the stable channel, but then there is a new one available by the time the
image finishes booting, so snapd would try to auto-refresh itself
simultaneously while we were trying to install the new snapd snap.

So download and unpack the snap and seed it into the image as local, unasserted
snap to prevent snapd from refreshing itself.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>