interfaces/apparmor: check for bpf capability support and generate snippet for s-c
When snap-confine runs on a cgroup v2 system, it will try to set up a device
filtering by installing a BPF program which uses BPF maps. This means that s-c
will perform a bpf() syscall. We should extend the snap-confine apparmor profile
to support this, but snapd runs on old systems where apparmor 2.x is still in
use, which does not understand `capability bpf` and fails when loading the
profiles.

For this reason, implement a workaround which probes the available parser
features and generates snippet for snap-confine. Existing `#include` entry in
the generic snap-confine profile will pick up the relevant bits automatically.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>