cmd/snap: improve snap disconnect arg parsing and err msg
snap disconnect's parsing would output an error message "invalid
value: (want snap:name or snap)" even though invoking it with just
the snap resulted in an error with a conflicting message.
This commit fixes the parsing for snap disconnect so it's done in a
single place and returns a single error message.