interfaces/microstack-support: set controlsDeviceCgroup to true
This will prevent us from generating any udev rules which in turn should 
entirely disable the enforcement of the device cgroup for processes in the 
snap.

This is justified by the snap already managing cgroups of its containers and 
VM's with the Delegate=true setting in serviceSnippets.

See also LP bug: https://bugs.launchpad.net/snapd/+bug/1892895 which is 
partially addressed by this commit, specifically for the microstack snap which
uses the microstack-support interface.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>