disks: rename SizeInSectors to SizeInBytes, adjust accordingly
To account for some of the confusion around sector sizes going forward, adjust
the methods to use bytes as the returned units directly and hide the complexity
inside the methods so that all the callers need to do is a type conversion to
quantity.Size and doesn't need to worry about the sector size.

Due to this, add an additional unit test about 4K sector size on GPT where we
need to call blockdev and sfdisk.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>