gadget: replace ondisk implementation with disks package, refactor part calcs
This enables us to write much simpler tests for complicated disk setups, which
will be critical for supporting multiple volume update situations, as otherwise
we would be left writing extremely complicated sfdisk and lsblk mock commands
which is no fun.

This also refactors the calculation of the size of partitions to use the more
natural and specific field "UsableSectorsEnd" which makes the code more 
readable and clear about what is going on without using the Size field, which
isn't quite what we want for the partition size calculation of expanded 
partitions.

Also delete the short-lived MockLsblkCommand in the gadgettest package, its
utility is no more.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>