osutil/disks: re-org methods for end of usable region, size information
Split off the LastUsableByte function into two methods, SizeInBytes and 
UsableSectorsEnd. The SizeInBytes function returns the most basic, commonly
understood size of a disk, including all regions, even those that are not 
usable by user partitions, etc.

The UsableSectorsEnd method however now essentially returns the last sector of
the disk that cannot be used, such that a partition which is specified as 
occupying [start, end) at the end of the disk can be calculated easily with
this method. 

This also refactors some of the tests to be more clear what they are testing
and test related methods in the same test.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>