many: quote constructed paths ending up into AppArmor rules
This helps ensuring that a path containing spaces or commas is properly
encoded into AppArmor profiles. Note that this is not a security fix:
one can still create a malicious path like

    source: path1" r, /etc/passwd rw, "path2

which would grant the snap read-write access to /etc/passwd.

This first patch, however, makes it easier to prevent this by simply
rejecting paths containing a `"` (should we decide to go down that
route; but in any case the presence of a `"` could be used to trigger a
more security-sensitive validation).