i/b/content: ensure that paths don't express AppArmor patterns
The content interface does not allow globbing in path names. Since the
paths are dumped verbatim into the apparmor profile, we must make sure
that they cannot be used to inject malicious rules into the profile.
Since we are already surrounding the path strings with quotes, all what
is left to do is to ensure that paths do not contain a `"` character.

However, while that alone would protect from rules injection, we'd still
be open to the risk of having a malformed path pattern cause a failure
in apparmor_parser, because quoted paths are still being interpreted as
apparmor glob patterns. A closing brace (`}`) not preceded by an opening
one, for example, would be detected as an error and our profile would
fail to parse. That's why are also excluding all the AppArmor globbing
characters.

We could also simply quote them, but as long as there isn't a
requirement to allow these characters, just blocking them seems safer.