tests: initialize CHANGE_ID in _wait_autorefresh
There was a failure in the snapd tests:
```
...
2021-11-02T16:55:11.8447301Z ++ /home/gopath/src/github.com/snapcore/snapd/tests/lib/tools/snapd-state wait-for-autorefresh 19
2021-11-02T16:55:11.8449962Z /home/gopath/src/github.com/snapcore/snapd/tests/lib/tools/snapd-state: line 92: [: : integer expression expected
...
```
and then a subsequent failure. Looking at snapd-state it seems like
the above failure will happen if _wait_autorefresh() times out. In
this case CHANGE_ID is never set and the line:
```
if [ "$LAST_CHANGE_ID" -eq "$CHANGE_ID" ]; then
```
errors because CHANGE_ID is never initialized and empty.

This commit initializes it which prevents this error.