osutil/disks: make PartitionType always lower case, add tests for this
We were missing tests for the codepath for DOS disks too, so add that in, and
adjust the check for the schema to check for lowercase "dos" which is what we
should always have (a followup commit will make this so).

Also the doc-comment was slightly misleading, we don't get "0x0C", we just get
"0C".

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>