Allow access to hunspell dictionaries for spellchecking
This commit is a step to address this bug:
https://bugzilla.mozilla.org/show_bug.cgi?id=1732755

Users should be able to install spellchecking dictionaries as needed
and share those also with the confined Firefox snap. If combined with
setting the DICPATH environment variable, this rule will prevent a
functional regression for Ubuntu users who switch from Firefox installed
as .deb to the Firefox snap.