osutil/disks: introduce RegisterDeviceMapperBackResolver + use for crypt-luks2
Add a new callback functionality for registering handlers which "un-map" a
device mapper node back to the original source for purposes of creating and/or
locating a disk device.

The CRYPT-LUKS2 based implementation currently supported is now implemented as
a handler this way, and this paves the way for additional handlers to be
implemented as well, specifically in other packages so that the osutil/disks
package can remain as generic as possible and easily imported from anywhere
without import cycles.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>