.github/workflows/test.yaml: set pipefail when running spread
This is to ensure when we do things like `spread ... | tee spread.log`, if
spread fails, the pipe command will fail as well which leads to the action
being correctly marked as failed.

Without this flag set, the tee command swallows all failures in the spread
command leading all checks to be marked as successful.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>