cmd/libsnap-confine-private: do not deny all devices when reusing the device cgroup
With device cgroup v1, when reusing the cgroup (i.e. opening with
SC_DEVICE_CGROUP_FROM_EXISTING flag), we should not deny all devices, as this
will negatively affect the processes that are in the group.

This code path was executed by snap-device-helper, so it is possible that when
processing of real events from device changes the group could have become
broken.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>