tests: ensure machine clock is ntp synced during prepare
We see failures in some tests now because the auto-refresh
timers wait for ntp to be available. For some reason not
all GCE machines sync their clocks. It looks a bit random
so to workaround this this commit explicitly enables the
ntp sync.

In addition it expands the debug for the two failing tests
so that if this is not enough we get better information what
is going on.