tests: disable flaky uc18 tests until systemd is fixed
Because of a bug/regression in systemd (LP:1949089) we see a bunch
of test failures on UC18. While the bug gets fixed we need to
disable the tests to avoid having unmergable PRs for reasons
outside of our control.

Once this is merged I will push a PR that reverts this commit so
that we can track the progress of the systemd fix.