o/devicestate: during remodel first check pending download tasks for snaps
When remodeling, it is possible that the new model will use a snap of the same
name, but from a different track/channel which may in fact carry different
content. Thus when creating a recovery system, make sure to use the 'new' snap
if it is part of the download tasks, rathe than first checking the local ones.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>