Add basic tag for tests
The idea of this change is to tag tests, those tags can be used then by
spread to define which tests have to be executed.

This first change includes the basic tests, which cover the basic
functionalities of snapd. Next step is to define which tests cover
alternative paths of the features, which cover errors, which are
regressions, etc.