systemd: add support for .target units
systemd/systemd.go: provides systemctl Status() support for socket,
mount, service and timer units.

- Add Status() support for .target units.

- Add unit test for .target unit

This is required to query the state of the unit before an
operation such as 'mask', 'unmask', 'start', 'stop', 'enable'
or 'disable' can be applied.

This patch is required in order to implement support for
controlling ctrl-alt-del behaviour, since the unit type is
'target'.

Signed-off-by: Fred Lotter <fred.lotter@canonical.com>