packaging: stop building with buildmode=pie and strip binary
This commit removes buildmode=pie because with go-1.13 the size
of the snapd snap increases by 10MB on amd64 and 9MB on arm64.

While PIE provides security benefits the trade-off between this
extra size is probably not worth it, especially considering that
some of our devices are fairly low end.

This commit also add `-s` to the `ldflags` to strip the binary.