interfaces/apparmor: Fix possible nil error report
In the case where apparmor_parser returns 0 but contains the output "parser
error" we would correctly signal an error back to the caller but this would
contain the printed representation of an error object which was actually
nil. In this case, we manually create an error which can be printed and
reported back to the caller.

Signed-off-by: Alex Murray <alex.murray@canonical.com>