interfaces/apparmor: Add backend test for apparmor_parser error
Older versions of apparmor_parser may actually exit with 0 status but fail
to parse the profile - this can be detected by looking for "parser error"
in the output from apparmor_parser - and the current apparmor backend code
correctly does this. So add a test to ensure we are testing this behaviour
going forward.

Signed-off-by: Alex Murray <alex.murray@canonical.com>