packaging/ubuntu, packaging/debian: depend on dbus-session-bus provider
On a cgroup v2 system, we rely on being able to set up a systemd scope for the
snap applications. This is done by communicating with a proper systemd instance
over a dbus bus, which for user session is the session bus.

This only came up when one of the intermediate dependencies of ubuntu-server,
had their dependencies updated and no longer requires session bus and we
observed that it was impossible to launch snap applications that required device
access on Ubuntu 21.10.

Related to:
- https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/1951491
- https://forum.snapcraft.io/t/cannot-launch-snap-applications-with-cgroup-v2/27700

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>