interfaces/udev: fix installing snaps inside lxd in 21.10
* interfaces/udev: fix installing snaps inside lxd in 21.10

In 21.10 a snap install inside an LXD container can trigger the an
error from `udevadm trigger`. This was reported in
https://github.com/lxc/lxd/issues/9526 and the advise from the
LXD folks is to ignore `udevadm trigger` exit-code errors. Only
systemd 248 started to report errors so they have always been
there but where just never reported before.

Plus as Ian put it so eloquently: We aren't really in the business
of reacting to any udevadm errors here, so there isn't really anything
we could do about errors that do crop up when we run this.
We could maybe try and detect that we are inside a container, and only
ignore udevadm errors in that case rather than always ignore udevadm
errors, but I'm not convinced that these new changes to systemd in
21.10 and newer which make udevadm return an error code won't also
some day make native hosts running new udevadm return an error code
"by default" on some folks' system in which case we would probably
have no option except to just remove that check for being in a
container anyways.

* udev: rework error handling (thanks to Pawel)

* udev: add some better comments (thanks to Maciej)