packaging/ubuntu-16.04: depend on dbus-session-bus if not on 16.04
The (default-)dbus-session-bus was introduced only in 2018, so it's available in
18.04 but not in earlier releases, so revert to not having a special dependency
on that release. Not having a session bus and lack of snap app tracking is not
fatal on 16.04, but is a must on 21.10 where a snap specific cgroup created by
dbus user instance is required for device access filtering.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>