packaging/ubuntu, packaging/debian: depend on dbus-session-bus provider (#11111)
* packaging/ubuntu, packaging/debian: depend on dbus-session-bus provider

On a cgroup v2 system, we rely on being able to set up a systemd scope for the
snap applications. This is done by communicating with a proper systemd instance
over a dbus bus, which for user session is the session bus.

This only came up when one of the intermediate dependencies of ubuntu-server,
had their dependencies updated and no longer requires session bus and we
observed that it was impossible to launch snap applications that required device
access on Ubuntu 21.10.

Related to:
- https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/1951491
- https://forum.snapcraft.io/t/cannot-launch-snap-applications-with-cgroup-v2/27700

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* spread: have ubuntu-core-18 be based on 18.04 images

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* packaging/ubuntu-16.04: depend on dbus-session-bus if not on 16.04

The (default-)dbus-session-bus was introduced only in 2018, so it's available in
18.04 but not in earlier releases, so revert to not having a special dependency
on that release. Not having a session bus and lack of snap app tracking is not
fatal on 16.04, but is a must on 21.10 where a snap specific cgroup created by
dbus user instance is required for device access filtering.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* tests/lib/pkgdb: dbus-user-session shall be pulled in through default-dbus-session-bus dependency

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* tests/lib/pkgdb: fix indentation

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* tests/main/lxd-snapfuse,lxd-try,lxd-mount-units: wait for container to be booted, run apt update before installing any packages

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* tests/main/lxd-try/task.yaml: fix typo

Co-authored-by: Ian Johnson <person.uwsome@gmail.com>