o/c/c/hostname: be more robust against future code changes
If more fields are added to the `fsOnlyContext` structure, or if it will
be initialized with an empty `rootDir` field, the code would take the
`else` path and directly write the hostname file, even though the path
using `hostnamectl` would work (and is arguably preferable).

At the moment it seems that the code in
overlord/configstate/configcore/handlers.go is always initializing a
`fsOnlyContext` structure with a non-empty `rootDir`, so this really
does not matter much for the time being.