snap: add support for `snap watch --last={revert,enable,disable,swtich}`
While looking at our code I noticed that we are a bit inconsistent
in our changeIDMixin about where we accept the "short" form for
a change and where we don't. E.g. for the change type `install-snap`
we support the short form `install` but for others like `revert-snap`
we do not.

This commit unifies this some more and fixes a bug in our testsuite
along the way where we use `snap watch --last=revert?` which does
not work today.