gadget/install/partition_test.go: rework tests, add non-adjacent test case
We don't need the /dev/node -> /dev/updated-node trick anymore since we don't
end up inspecting the disk twice, but since we do end up modifying the
OnDiskVolume, we should be checking that after RemoveCreatedPartitions is
called.

Also add a unit test where the structure that is deleted is not the last one to
show that partitions which we don't delete can be non-adjacent and it still
works. Thanks to Samuele for the suggestion.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>