Skip installing systemd-timesyncd in debian 10
This packages is having dependency issues to be installed in debian 10,
so images cannot be updated correctly on spread-images project.

The error is:

The following packages have unmet dependencies:
systemd-timesyncd : Depends: systemd (= 247.3-6~bpo10+1) but
241-7~deb10u8 is to be installed
Breaks: systemd (< 245.4-2~) but 241-7~deb10u8 is
to be installed
E: Unable to correct problems, you have held broken packages.
quiet: end of output.