gadget/update: rename DiskVolume...ForDevice -> DiskTraitsFromDeviceAndValidate
Also adjust the loop to ensure that the mapped structure partitions we create
come from the laid out volume so we can easily double check there are no
partitions sneaking into the device traits map that were not in the
gadget.yaml.

Also adjust some comments.

Thanks to Alberto and Samuele for the suggestions.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>