cmd/snap-device-helper: bring back the device type identification behavior, but for remove action fallback only
The old shell-based snap-device-helper would identify the devices based on
pattern matching the device path. Specifically any device matching */block/*
or */nvme/nvme*/nvme*n* patterns was considered a block device, while all other
devices were considered to be char devices [1].

Bring back the old behavior, but only in a fallback path of the remove action.
Since when devices are removed, it is possible that the symlinks are already
gone by the time we inspect entries in /sys/devices/. However, we would still
like to remove the device from the cgroup.

1. https://github.com/snapcore/snapd/blob/cd99de559403e3ffd44f09811427fc873c7e85a3/cmd/snap-confine/snap-device-helper#L55-L68

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>