tests: fix remodel-kernel test when running on external devices
Save/restore the tracking channel of the kernel in the test to
fix a failure when this test runs on e.g. external devices that
have models where the tracking channel of the kernel does not
match the "$KERNEL_CHANNEL" we use when we build the images on
our own.

While at it I removed a bunch of hardcoded "pc-kernel" strings.