usersession/userd: query xdg-mime to check for fallback handlers of a given scheme
Most of the handling is of xdg-open requests from the snaps id done by
xdg-desktop-protal now, the handler in userd remains a fallback for scenarios
where the desktop setup is incomplete. The code of io.snapcraft.OpenURL()
handler would only allow a handful of schemes to be passed to xdg-open on the
host side. However, updating the list of schemes manually has proven be to
unmaintainable, and got filled with various vendor specific entries.

In https://github.com/snapcore/snapd/pull/7731#issuecomment-585721100 an
approach was proposed to use xdg-query to find out whether there is handler for
given scheme on the host side, and if so allow the URL to be passed to xdg-open,
which is implemented in this patch.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>