mkversion.sh: do full opts parsing, add --ignore-debian-rules-changes
Re-implement options parsing using bash case/esac flow with shift, etc. to be
able to support more explicit options and also display a help message with
invalid options or --help option.

Also add a new option behavior with --ignore-debian-rules-changes which
specifically will treat the specific change that is caused by the debian/rules
file when handling libseccomp on non-trusty platforms as not being "dirty". If
there are other changes in addition to those changes, then the tree is treated
as being dirty, otherwise if only those changes are present and the option is
specified we treat it as if the tree was pristine for the purposes of setting
the version number. This is implemented through two patches, one which is the
exact expected change that debian/rules file should be performing, and one
which is the exact revert of those changes. This allows us to detect what
changes are present and undo any changes we do locally.

Finally use the new options as relevant across the packaging scripts.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>