github: test for "dirty" in usr/lib/snapd/info
During the 2.54 cycle we fixed the "dirty" version during the
snapcraft build. Unfortunately the "dirty" tag creeped into
our version information that was generated during the package
build and resulted in a "dirty" version in `usr/lib/snapd/info`
and in the auto-generated `snapdtools/version_generated.go`.

To detect this as well, this commit adds an extra check to the
workflow for this.