tests: make `distro_install_package` robust in `if` context
When using a shell function inside an `if` context the `set -e`
setting is disabled. This means that `distro_install_package`
when used in a `if distro_install_package` context will not
fail if `apt` (or any of the others) fails. This commit add
manual return value checking to ensure this works correctly.