gadget/update.go: handle YAML's with bare structures
We cannot measure traits for bare structures that are not in the partition
table since by definition these structures are not mentioned at all in the
partition table and thus Linux knows nothing about them.

Add a unit test for this and rename some of the previous structures which I
erroneously thought were bare structures but actually were real partitions,
just ones without filesystems.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>