tests: fix parallel-install-basic on external UC16 devices
We saw some test failures running this test on UC16 external
devices. It turns out the failures are a side-effect of the
way `tests.session exec` works on UC16. Here the `/snap/bin`
directory is not in `PATH`.

The other places that use `tests.session exec` workaround this
by prepending `snap run`. This commit does the same for
the `parallel-install-basic` test (thanks to Maciej for the
suggestion).